Adding openpgp.worker this improves the Zimlets performance in Firefox (generate keys no longer hangs).
Unfortunately we cannot call openpgp.initWorker in the Zimlet constructor or prototype.init as this
will fail. (Zimbra webclient startup takes some time and openpgp.js is not loaded in time for this)
Therefore we call openpgp.initWorker in all other functions that use openpgp. This is dirty, but don't
see another way.

Fixing bugs in the generate keypair window in case the user does not enter user id or passphrase.